+++ src/r-wrappers.h
@@ -81,4 +81,4 @@ SEXP SXP_save_chrsplit_block_effects(SEXP exd, SEXP filename, SEXP nslices, SEXP
 /*--------------------------------Deletors------------------------------------*/
 SEXP clear_simdata(SEXP exd);
 void SXP_delete_simdata(SEXP sd);
-SEXP SXP_delete_group(SEXP exd, SEXP group);
+SEXP SXP_delete_group(SEXP exd, SEXP len, SEXP groups);