+++ NEWS.md
@@ -1,5 +1,7 @@
+# genomicSimulation dev
 - Fix segfault in select.by.gebv when trying to select more individuals than exist in the group (e.g. asking for the best 5 members of a group of 2). Now, it just moves all group members to the new selected group, and doesn't worry about the missing requested remainder.
 - adding new group splitter functions. Added tests for them.
+- `delete.group` can now be passed a vector of group ids to delete from memory. Previously it could only delete a single group at a time. 
 
 # genomicSimulation 0.2